+++ packages/ui/components/ui/dutch-auction-component/progress-bar/indicator/get-progress.test.ts
