+++ packages/ui/components/ui/dutch-auction-component/progress-bar/indicator/index.tsx
@@ -1,6 +1,6 @@
 import { DutchAuction } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/component/auction/v1/auction_pb';
 import { CircleArrowRight, CircleCheck, CircleX } from 'lucide-react';
-import { getProgress } from '../get-progress';
+import { getProgress } from './get-progress';
 import { getDescription } from '@penumbra-zone/getters/dutch-auction';
 import { isZero } from '@penumbra-zone/types/amount';
